// service-plat/src/types/JobPost.d.ts
export interface Jobs {
    user_id: string;
    name: string;
    cover: string;
    logo_url: string;
    image_url: string;
    product_count: number;
    is_following: boolean;
    follower_count: number;
    activity_sector: string;
    description: string | null;
    // pays: string;
    // ville: string;
    // rating: number;
    // activite: string;
  }
  